+++ VSC Extension/src/server.js
@@ -1,24 +1,17 @@
 const http = require('http')
 const vscode = require('vscode')
 const path = require('path')
-const ffi = require('ffi-napi')
 const config = require('./config/settings.js')
 const website = require('./config/website.js')
 const events = require('./events')
 const files = require('./files')
 const twoWaySync = require('./twoWaySync')
 const apiDump = require('./config/apiDump')
-const wcpp = require('wcpp')
+
+//@ts-ignore
+const winuser = require('./utils/winuser')
 
 const URL = 'http://$host:$port/'
-const user32 = new ffi.Library('user32', {
-    'GetForegroundWindow': ['long', []],
-    'ShowWindow': ['bool', ['long', 'int']],
-    'SetForegroundWindow': ['bool', ['long']],
-
-    'GetAsyncKeyState': ['short', ['int']],
-    'keybd_event': ['void', ['long', 'long', 'long', 'long']]
-})
 
 let server = http.createServer(requestListener)
 let lastSync = Date.now()
@@ -26,8 +19,6 @@ let sockets = new Set() //Temp fix for forcing server to stop (until Electron ad
 let isConnected = false
 let requestsLeft = 0
 let chunks = []
-let vscWindow
-let studioWindow
 
 let syncCount = 5
 let uptime = 0
@@ -94,10 +85,6 @@ function requestListener(request, response) {
             if (!isConnected) {
                 lastSync = Date.now()
                 isConnected = true
-            }
-
-            if (!studioWindow) {
-                studioWindow = user32.GetForegroundWindow() //unreliable af, will change once I rewrite some of the code in C++
             }
             break
         case 'getState':
@@ -214,56 +201,14 @@ function openFile(file) {
 
     vscode.workspace.openTextDocument(file).then(file => {
         vscode.window.showTextDocument(file, {preview: config.openInPreview})
-
-        if (vscWindow) {
-            let pressed = false
-
-            events.queue.push({Action: 'closeFile'})
-
-            if ((user32.GetAsyncKeyState(0x12) & 0x8000) == 0) {
-                user32.keybd_event(0x12, 0, 0x0001 | 0, 0)
-                pressed = true
-            }
-
-            user32.ShowWindow(vscWindow, 3)
-            user32.SetForegroundWindow(vscWindow)
-
-            if (pressed) {
-                user32.keybd_event(0x12, 0, 0x0001 | 0x0002, 0)
-            }
-        }
+        events.queue.push({Action: 'closeFile'})
+        winuser.showVSC()
     }).then(undefined, () => {})
-}
-
-function debug() {
-    if (studioWindow) {
-        let pressed = false
-
-        if ((user32.GetAsyncKeyState(0x12) & 0x8000) == 0) {
-            user32.keybd_event(0x12, 0, 0x0001 | 0, 0)
-            pressed = true
-        }
-
-        user32.ShowWindow(studioWindow, 3)
-        user32.SetForegroundWindow(studioWindow)
-
-        if (pressed) {
-            user32.keybd_event(0x12, 0, 0x0001 | 0x0002, 0)
-        }
-
-        setTimeout(() => {
-            user32.keybd_event(0x74, 0, 0x0001 | 0, 0)
-            setTimeout(() => {
-                user32.keybd_event(0x74, 0, 0x0001 | 0x0002, 0)
-            }, 100); 
-        }, 100);
-    }
 }
 
 module.exports = {
     run,
-    stop,
-    debug
+    stop
 }
 
 server.on('connection', (socket) => {
@@ -273,15 +218,3 @@ server.on('connection', (socket) => {
         sockets.delete(socket);
     })
 })
-
-if (vscode.window.state.focused) {
-    vscWindow = user32.GetForegroundWindow()
-}
-else {
-    let watcher = vscode.window.onDidChangeWindowState((state) => {
-        if (state.focused) {
-            vscWindow = user32.GetForegroundWindow()
-            watcher.dispose()
-        }
-    })
-}